+++ plugin/path_role_set_secrets.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package gcpsecrets
 
 import (
@@ -18,15 +21,18 @@ func fieldSchemaRoleSetServiceAccountKey() map[string]*framework.FieldSchema {
 			Type:        framework.TypeString,
 			Description: fmt.Sprintf(`Private key algorithm for service account key - defaults to %s"`, keyAlgorithmRSA2k),
 			Default:     keyAlgorithmRSA2k,
+			Query:       true,
 		},
 		"key_type": {
 			Type:        framework.TypeString,
 			Description: fmt.Sprintf(`Private key type for service account key - defaults to %s"`, privateKeyTypeJson),
 			Default:     privateKeyTypeJson,
+			Query:       true,
 		},
 		"ttl": {
 			Type:        framework.TypeDurationSecond,
 			Description: "Lifetime of the service account key",
+			Query:       true,
 		},
 	}
 }
@@ -43,12 +49,26 @@ func fieldSchemaRoleSetAccessToken() map[string]*framework.FieldSchema {
 
 func pathRoleSetSecretServiceAccountKey(b *backend) *framework.Path {
 	return &framework.Path{
-		Pattern:        fmt.Sprintf("roleset/%s/key", framework.GenericNameRegex("roleset")),
+		Pattern: fmt.Sprintf("roleset/%s/key", framework.GenericNameRegex("roleset")),
+		DisplayAttrs: &framework.DisplayAttributes{
+			OperationPrefix: operationPrefixGoogleCloud,
+			OperationVerb:   "generate",
+		},
 		Fields:         fieldSchemaRoleSetServiceAccountKey(),
 		ExistenceCheck: b.pathRoleSetExistenceCheck("roleset"),
 		Operations: map[logical.Operation]framework.OperationHandler{
-			logical.ReadOperation:   &framework.PathOperation{Callback: b.pathRoleSetSecretKey},
-			logical.UpdateOperation: &framework.PathOperation{Callback: b.pathRoleSetSecretKey},
+			logical.ReadOperation: &framework.PathOperation{
+				Callback: b.pathRoleSetSecretKey,
+				DisplayAttrs: &framework.DisplayAttributes{
+					OperationSuffix: "roleset-key2",
+				},
+			},
+			logical.UpdateOperation: &framework.PathOperation{
+				Callback: b.pathRoleSetSecretKey,
+				DisplayAttrs: &framework.DisplayAttributes{
+					OperationSuffix: "roleset-key",
+				},
+			},
 		},
 		HelpSynopsis:    pathServiceAccountKeySyn,
 		HelpDescription: pathServiceAccountKeyDesc,
@@ -57,13 +77,27 @@ func pathRoleSetSecretServiceAccountKey(b *backend) *framework.Path {
 
 func deprecatedPathRoleSetSecretServiceAccountKey(b *backend) *framework.Path {
 	return &framework.Path{
-		Pattern:        fmt.Sprintf("key/%s", framework.GenericNameRegex("roleset")),
+		Pattern: fmt.Sprintf("key/%s", framework.GenericNameRegex("roleset")),
+		DisplayAttrs: &framework.DisplayAttributes{
+			OperationPrefix: operationPrefixGoogleCloud,
+			OperationVerb:   "generate",
+		},
 		Deprecated:     true,
 		Fields:         fieldSchemaRoleSetServiceAccountKey(),
 		ExistenceCheck: b.pathRoleSetExistenceCheck("roleset"),
 		Operations: map[logical.Operation]framework.OperationHandler{
-			logical.ReadOperation:   &framework.PathOperation{Callback: b.pathRoleSetSecretKey},
-			logical.UpdateOperation: &framework.PathOperation{Callback: b.pathRoleSetSecretKey},
+			logical.ReadOperation: &framework.PathOperation{
+				Callback: b.pathRoleSetSecretKey,
+				DisplayAttrs: &framework.DisplayAttributes{
+					OperationSuffix: "roleset-key4",
+				},
+			},
+			logical.UpdateOperation: &framework.PathOperation{
+				Callback: b.pathRoleSetSecretKey,
+				DisplayAttrs: &framework.DisplayAttributes{
+					OperationSuffix: "roleset-key3",
+				},
+			},
 		},
 		HelpSynopsis:    pathServiceAccountKeySyn,
 		HelpDescription: pathServiceAccountKeyDesc,
@@ -72,12 +106,26 @@ func deprecatedPathRoleSetSecretServiceAccountKey(b *backend) *framework.Path {
 
 func pathRoleSetSecretAccessToken(b *backend) *framework.Path {
 	return &framework.Path{
-		Pattern:        fmt.Sprintf("roleset/%s/token", framework.GenericNameRegex("roleset")),
+		Pattern: fmt.Sprintf("roleset/%s/token", framework.GenericNameRegex("roleset")),
+		DisplayAttrs: &framework.DisplayAttributes{
+			OperationPrefix: operationPrefixGoogleCloud,
+			OperationVerb:   "generate",
+		},
 		Fields:         fieldSchemaRoleSetAccessToken(),
 		ExistenceCheck: b.pathRoleSetExistenceCheck("roleset"),
 		Operations: map[logical.Operation]framework.OperationHandler{
-			logical.ReadOperation:   &framework.PathOperation{Callback: b.pathRoleSetSecretAccessToken},
-			logical.UpdateOperation: &framework.PathOperation{Callback: b.pathRoleSetSecretAccessToken},
+			logical.ReadOperation: &framework.PathOperation{
+				Callback: b.pathRoleSetSecretAccessToken,
+				DisplayAttrs: &framework.DisplayAttributes{
+					OperationSuffix: "roleset-access-token2",
+				},
+			},
+			logical.UpdateOperation: &framework.PathOperation{
+				Callback: b.pathRoleSetSecretAccessToken,
+				DisplayAttrs: &framework.DisplayAttributes{
+					OperationSuffix: "roleset-access-token",
+				},
+			},
 		},
 		HelpSynopsis:    pathTokenHelpSyn,
 		HelpDescription: pathTokenHelpDesc,
@@ -86,13 +134,27 @@ func pathRoleSetSecretAccessToken(b *backend) *framework.Path {
 
 func deprecatedPathRoleSetSecretAccessToken(b *backend) *framework.Path {
 	return &framework.Path{
-		Pattern:        fmt.Sprintf("token/%s", framework.GenericNameRegex("roleset")),
+		Pattern: fmt.Sprintf("token/%s", framework.GenericNameRegex("roleset")),
+		DisplayAttrs: &framework.DisplayAttributes{
+			OperationPrefix: operationPrefixGoogleCloud,
+			OperationVerb:   "generate",
+		},
 		Deprecated:     true,
 		Fields:         fieldSchemaRoleSetAccessToken(),
 		ExistenceCheck: b.pathRoleSetExistenceCheck("roleset"),
 		Operations: map[logical.Operation]framework.OperationHandler{
-			logical.ReadOperation:   &framework.PathOperation{Callback: b.pathRoleSetSecretAccessToken},
-			logical.UpdateOperation: &framework.PathOperation{Callback: b.pathRoleSetSecretAccessToken},
+			logical.ReadOperation: &framework.PathOperation{
+				Callback: b.pathRoleSetSecretAccessToken,
+				DisplayAttrs: &framework.DisplayAttributes{
+					OperationSuffix: "roleset-access-token4",
+				},
+			},
+			logical.UpdateOperation: &framework.PathOperation{
+				Callback: b.pathRoleSetSecretAccessToken,
+				DisplayAttrs: &framework.DisplayAttributes{
+					OperationSuffix: "roleset-access-token3",
+				},
+			},
 		},
 		HelpSynopsis:    pathTokenHelpSyn,
 		HelpDescription: pathTokenHelpDesc,