+++ plugin/rollback.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package gcpsecrets
 
 import (
@@ -222,7 +225,8 @@ func (b *backend) serviceAccountPolicyRollback(ctx context.Context, req *logical
 		return err
 	}
 
-	apiHandle := iamutil.GetApiHandle(httpC, useragent.String())
+	apiHandle := iamutil.GetApiHandle(httpC, useragent.PluginString(b.pluginEnv,
+		userAgentPluginName))
 	p, err := r.GetIamPolicy(ctx, apiHandle)
 	if err != nil {
 		if isGoogleAccountNotFoundErr(err) || isGoogleAccountUnauthorizedErr(err) {
@@ -284,7 +288,8 @@ func (b *backend) serviceAccountPolicyDiffRollback(ctx context.Context, req *log
 		return err
 	}
 
-	apiHandle := iamutil.GetApiHandle(httpC, useragent.String())
+	apiHandle := iamutil.GetApiHandle(httpC, useragent.PluginString(b.pluginEnv,
+		userAgentPluginName))
 	p, err := r.GetIamPolicy(ctx, apiHandle)
 	if err != nil {
 		return err